Code

+++ AuctionExpress.Service/TransactionService.cs
@@ -33,8 +33,29 @@ namespace AuctionExpress.Service
             }
         }
 
-        //GET Transaction Currently returning all transactions where the signed in user is the winner
+        //GET All Transactions
+        public IEnumerable<TransactionListItem> GetAllTransactions()
+        {
+            using (var ctx = new ApplicationDbContext())
+            {
+                var query =
+                    ctx
+                    .Transaction
+                    .Select(e => new TransactionListItem
+                    {
+                        TransactionId = e.TransactionId,
+                        ProductId = e.ProductId,
+                        ProductName = e.TransactionProduct.ProductName,
+                        SellerId = e.TransactionProduct.ProductSeller,
+                        IsPaid = e.IsPaid,
+                        PaymentDate = e.PaymentDate
+                    }
+                    );
+                return query.ToList();
+            }
+        }
 
+        //GET Transaction Currently returning all transactions where the signed in user is the winner
         public IEnumerable<TransactionListItem> GetTransactions()
         {
             using (var ctx = new ApplicationDbContext())
@@ -85,12 +106,13 @@ namespace AuctionExpress.Service
         }
 
 
-        //Update Transaction <--Update IsPaid, currently Sellers allowed to edit
+        //Update Transaction <--Updates IsPaid, currently Sellers allowed to edit
 
         public bool UpdateTransaction(TransactionEdit model)
         {
             using (var ctx = new ApplicationDbContext())
             {
+
                 var entity =
                     ctx
                     .Transaction
@@ -105,7 +127,7 @@ namespace AuctionExpress.Service
             }
         }
 
-        public bool DeleteTransaction(int transactionId)
+        public string DeleteTransaction(int transactionId)
         {
             using (var ctx = new ApplicationDbContext())
             {
@@ -113,11 +135,25 @@ namespace AuctionExpress.Service
                     ctx
                         .Transaction
                         .Single(e => e.TransactionId == transactionId);
+                if (entity.TransactionProduct.ProductSeller == _userId.ToString())
+                {
+                    try
+                    {
+                        ctx.Transaction.Remove(entity);
+                        ctx.SaveChanges();
 
-                ctx.Transaction.Remove(entity);
+                        return "Transaction successfully deleted";
+                    }
+                    catch (Exception e)
+                    {
+                        return e.Message;
+                    }
+                }
 
-                return ctx.SaveChanges() == 1;
+                return "User unauthorized to delete this transaction.";
             }
+
         }
     }
+
 }
